+++ MatrixSDK/MXRestClient.h
@@ -32,7 +32,7 @@
 #import "MXEventTimeline.h"
 #import "MXJSONModels.h"
 #import "MXFilterJSONModel.h"
-
+#import "MXMatrixVersions.h"
 
 #pragma mark - Constants definitions
 /**
@@ -173,6 +173,21 @@ typedef enum : NSUInteger
 -(id)initWithCredentials:(MXCredentials*)credentials andOnUnrecognizedCertificateBlock:(MXHTTPClientOnUnrecognizedCertificate)onUnrecognizedCertBlock NS_REFINED_FOR_SWIFT;
 
 - (void)close;
+
+
+#pragma mark - Server administration
+/**
+ Gets the versions of the specification supported by the server.
+
+ @param success A block object called when the operation succeeds. It provides
+                the supported spec versions.
+ @param failure A block object called when the operation fails.
+
+ @return a MXHTTPOperation instance.
+ */
+- (MXHTTPOperation*)supportedMatrixVersions:(void (^)(MXMatrixVersions *matrixVersions))success
+                        failure:(void (^)(NSError *error))failure;
+
 
 #pragma mark - Registration operations
 /**