+++ MatrixSDKTests/MXRestClientNoAuthAPITests.m
@@ -84,6 +84,55 @@
 }
 
 
+#pragma mark - Server administration
+- (void)testSupportedMatrixVersions
+{
+    XCTestExpectation *expectation = [self expectationWithDescription:@"asyncTest"];
+
+    [mxRestClient supportedMatrixVersions:^(MXMatrixVersions *matrixVersions) {
+
+        XCTAssertNotNil(matrixVersions);
+        XCTAssertNotNil(matrixVersions.versions);
+
+        // Check supported spec version at the time of writing this test
+        XCTAssert([matrixVersions.versions containsObject:@"r0.0.1"]);
+        XCTAssert([matrixVersions.versions containsObject:@"r0.1.0"]);
+        XCTAssert([matrixVersions.versions containsObject:@"r0.2.0"]);
+        XCTAssert([matrixVersions.versions containsObject:@"r0.3.0"]);
+
+        [expectation fulfill];
+
+    } failure:^(NSError *error) {
+        XCTFail(@"The request should not fail - NSError: %@", error);
+        [expectation fulfill];
+    }];
+
+    [self waitForExpectationsWithTimeout:10 handler:nil];
+}
+
+// At the time of introducing this test, MXMatrixVersions.supportLazyLoadMembers
+// was stored in MXMatrixVersions.unstableFeatures.
+// Make sure that, in future versions of the spec, supportLazyLoadMembers is still YES
+// wherever it will be stored.
+- (void)testSupportedMatrixVersionsSupportLazyLoadMembers
+{
+    XCTestExpectation *expectation = [self expectationWithDescription:@"asyncTest"];
+
+    [mxRestClient supportedMatrixVersions:^(MXMatrixVersions *matrixVersions) {
+
+        XCTAssert(matrixVersions.supportLazyLoadMembers);
+
+        [expectation fulfill];
+
+    } failure:^(NSError *error) {
+        XCTFail(@"The request should not fail - NSError: %@", error);
+        [expectation fulfill];
+    }];
+
+    [self waitForExpectationsWithTimeout:10 handler:nil];
+}
+
+
 #pragma mark - Registration operations
 - (void)testGetRegisterSession
 {